+++ service/position.go
@@ -29,6 +29,7 @@ type PositionStore interface {
 	GetByMarketAndParty(ctx context.Context, marketID entities.MarketID, partyID entities.PartyID) (entities.Position, error)
 	GetByMarket(ctx context.Context, marketID entities.MarketID) ([]entities.Position, error)
 	GetByParty(ctx context.Context, partyID entities.PartyID) ([]entities.Position, error)
+	GetByPartyConnection(ctx context.Context, partyID entities.PartyID, marketID entities.MarketID, pagination entities.CursorPagination) ([]entities.Position, entities.PageInfo, error)
 	GetAll(ctx context.Context) ([]entities.Position, error)
 }
 
@@ -101,6 +102,10 @@ func (p *Position) GetByMarket(ctx context.Context, marketID entities.MarketID) 
 
 func (p *Position) GetByParty(ctx context.Context, partyID entities.PartyID) ([]entities.Position, error) {
 	return p.store.GetByParty(ctx, partyID)
+}
+
+func (p *Position) GetByPartyConnection(ctx context.Context, partyID entities.PartyID, marketID entities.MarketID, pagination entities.CursorPagination) ([]entities.Position, entities.PageInfo, error) {
+	return p.store.GetByPartyConnection(ctx, partyID, marketID, pagination)
 }
 
 func (p *Position) GetAll(ctx context.Context) ([]entities.Position, error) {